Chore: entity type change #2

See merge request s09-bigdata-dist-sub2/S09P22E204!4

+++ backend/moa/src/main/java/com/ssafy/moa/api/entity/key/KoreanKey.java
@@ -14,13 +14,13 @@ import java.io.Serializable;
 @NoArgsConstructor
 @EqualsAndHashCode
 public class KoreanKey implements Serializable {
-    private int koreanId;
+    private Long koreanId;
 
     @OneToOne
     private NationCode koreanNationCode;
 
     @Builder
-    public KoreanKey(int koreanId, NationCode koreanNationCode) {
+    public KoreanKey(long koreanId, NationCode koreanNationCode) {
         this.koreanId = koreanId;
         this.koreanNationCode = koreanNationCode;
     }
